refactor overview and settings pages into components
